Move `ExperimentPackage` to `execution`
